style(mini_cd): ファイル分割
