+++ includes/builtin.h
@@ -6,7 +6,7 @@
 /*   By: totaisei <totaisei@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/12 20:07:55 by nfukada           #+#    #+#             */
-/*   Updated: 2021/03/05 11:32:00 by totaisei         ###   ########.fr       */
+/*   Updated: 2021/03/05 14:52:22 by totaisei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,12 +15,13 @@
 
 # include "libft.h"
 
-int		is_builtin(char **args);
-int		exec_echo(char **args);
-int		exec_exit(char **args);
-int		exec_builtin(char **args);
-int		exec_cd(char **args);
-int		exec_pwd(char **args);
-void	bind_pwd_valie(void);
+int			is_builtin(char **args);
+int			exec_echo(char **args);
+int			exec_exit(char **args);
+int			exec_builtin(char **args);
+int			exec_cd(char **args);
+int			exec_pwd(char **args);
+void		bind_pwd_value(void);
+t_bool		try_change_dir(const char *destination);
 
 #endif
